+++ pyproject.toml
@@ -99,10 +99,9 @@ exclude_lines = [
 ]
 
 [project.scripts]
-h3xrecon-dataprocessor = "h3xrecon.server.dataprocessor:run"
-h3xrecon-jobprocessor = "h3xrecon.server.jobprocessor:run"
-h3xrecon-worker = "h3xrecon.worker.worker:run"
-h3xrecon-client = "h3xrecon.client.client:main"
+h3xrecon-data = "h3xrecon.workers.data:run"
+h3xrecon-parsing = "h3xrecon.workers.parsing:run"
+h3xrecon-recon = "h3xrecon.workers.recon:run"
 
 [tool.coverage.paths]
 h3xrecon = ["src/h3xrecon", "*/h3xrecon/src/h3xrecon"]
