+++ src/h3xrecon/__about__.py
@@ -2,4 +2,4 @@
 #
 # SPDX-License-Identifier: MIT
 
-__version__ = "0.6.0"
+__version__ = "0.6.0.dev6"
