--- src/h3xrecon/__init__.py
@@ -1,13 +1,0 @@
-# # SPDX-FileCopyrightText: 2024-present h3xit <h3xit@protonmail.com>
-# #
-# # SPDX-License-Identifier: MIT
-
-# from .server.jobprocessor import JobProcessor
-# from .server.dataprocessor import DataProcessor
-# from .worker.worker import Worker
-
-# __all__ = [
-#     'JobProcessor',
-#     'DataProcessor',
-#     'Worker',
-# ]
